change path to module since it loads modules now, minor tweaks to hooks
